Adds Warframe inventory integration (WIP)
Implements basic Warframe inventory data loading
and display to show player riven data.
Includes UI components for displaying and filtering veiled rivens.
Adds API endpoints for accessing inventory data.

Removes outdated Alecaframe client modules
Adds AES & CBC encryption crate
Adds settings for inventory data path
Adds `wf_inventory` command group
Updates existing dependencies
Adds new locales

+++ src-tauri/qf_api/src/types/decrypt_keys.rs
@@ -3,8 +3,8 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct DecryptKeys {
     #[serde(rename = "key")]
-    pub key: Vec<i64>,
+    pub key: Vec<u8>,
 
     #[serde(rename = "iv")]
-    pub iv: Vec<i64>,
+    pub iv: Vec<u8>,
 }
